(new-component) Merge the stackblitz instructions
I understand where you're going with seperating the stackblitz instructions.
To have a clear distinction of a cli and stackblitz path.

In this particular case, though, the main thread already mentions stackblitz and works from it.

It was confusing to click the instructions, click the link back.
And when i click next, it brings back to the same instructions.

In my opinion, this makes it easier to follow for someone who is still trying to juggle so much information.